api/gateway_api/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
requests>=2.31.0
mlflow>=2.10.0
pydantic>=2.0
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart
